19. 파일 수정 추가 - 버그 수정
 - 게시글 수정 클릭시  기존의 파일들이 보이는 상태에서 새로운 파일 추가
 - 저장 버튼 클릭시 기존의 파일들을 DB에서 삭제되고, 서버에서도 삭제 그리고  새로운 파일로 대체

  db경로는 삭제되지만 서버에 파일이 남아잏는 현상 해결!

+++ src/main/java/org/zerock/service/Impl/BoardServiceImpl.java
@@ -11,7 +11,8 @@ import org.zerock.domain.BoardVO;
 import org.zerock.domain.Criteria;
 import org.zerock.domain.SearchCriteria;
 import org.zerock.persistence.BoardDAO;
-import  org.zerock.service.BoardService;
+import org.zerock.service.BoardService;
+import org.zerock.util.UploadFileUtilsForBBS;
 
 @Service
 public class BoardServiceImpl implements BoardService {
@@ -42,19 +43,27 @@ public class BoardServiceImpl implements BoardService {
 
 	@Transactional
 	@Override
-	public void modify(BoardVO board) throws Exception {
+	public void modify(BoardVO board,String uploadPath) throws Exception {
+		UploadFileUtilsForBBS util = new UploadFileUtilsForBBS();
+		
 		dao.update(board);
 		Integer bno = board.getBno();
-		dao.deleteAttach(bno);
-		String[] files = board.getFiles();
+		List<String> fileNameList = dao.getAttach(bno);
 		
-		if(files == null) {
-			return;
+		if(util.deleteAllFiles(fileNameList,uploadPath)) {
+			dao.deleteAttach(bno);
+			String[] files = board.getFiles();
+			
+			if(files == null) {
+				return;
+			}
+			
+			for(String fileName : files){
+				dao.replaceAttach(fileName, bno);
+			}
 		}
 		
-		for(String fileName : files){
-			dao.replaceAttach(fileName, bno);
-		}
+		
 		
 	}
